# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: SKN API
  version: 1.0.0-hackathon
  description: |
    React 모바일 앱과 Spring Boot API가 실제로 사용하는 계약이다.
    인증은 HttpOnly 세션 쿠키를 사용하고 OpenAI API 키는 서버에서만 읽는다.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
  /auth/signup:
    post:
      tags: [Auth]
      security: []
      summary: 아이디와 비밀번호로 가입하고 세션을 시작한다
      operationId: signUp
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SignUpRequest' }
      responses:
        '201': { description: 가입됨, content: { application/json: { schema: { $ref: '#/components/schemas/Auth' } } } }
        '409': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: 아이디와 비밀번호로 로그인한다
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200': { description: 로그인됨, content: { application/json: { schema: { $ref: '#/components/schemas/Auth' } } } }
        '422': { $ref: '#/components/responses/Problem' }
...
  /auth/quick-accounts:
    get:
      tags: [Auth, Demo]
      security: []
      summary: 테스트 하니스에서 바로 접속할 계정을 조회한다
      description: TEST_HARNESS_ENABLED=true인 시연 환경에서만 제공한다. 회원가입한 일반 계정도 자동으로 포함된다.
      operationId: listQuickAccounts
      responses:
        '200': { description: 빠른 로그인 계정, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/QuickAccount' } } } } }
        '404': { $ref: '#/components/responses/Problem' }
  /auth/quick-login/{username}:
    post:
      tags: [Auth, Demo]
      security: []
      summary: 테스트 계정으로 세션을 바꾸다
      description: TEST_HARNESS_ENABLED=true인 시연 환경에서만 제공하며, 데모 계정은 별도 데모 진입 API를 사용한다.
      operationId: quickLogin
      parameters:
        - { name: username, in: path, required: true, schema: { type: string, pattern: '^[a-z0-9_]{4,24}$' } }
      responses:
        '200': { description: 계정 전환됨, content: { application/json: { schema: { $ref: '#/components/schemas/Auth' } } } }
        '404': { $ref: '#/components/responses/Problem' }
  /auth/me:
    get:
      tags: [Auth]
      summary: 현재 세션 계정을 조회한다
      operationId: getMe
      responses:
        '200': { description: 현재 계정, content: { application/json: { schema: { $ref: '#/components/schemas/Auth' } } } }
        '401': { $ref: '#/components/responses/Problem' }
    delete:
      tags: [Auth]
      summary: 현재 계정과 연결된 개인 기록을 삭제한다
      operationId: deleteAccount
      responses:
        '204': { description: 삭제됨 }
        '401': { $ref: '#/components/responses/Problem' }
        '403': { $ref: '#/components/responses/Problem' }
...
  /auth/onboarding:
    post:
      tags: [Auth]
      summary: 선택한 화장품을 저장하고 온보딩을 마친다
      description: PRODUCT를 고르면 focusProductId의 사용 경험까지 멱등적으로 시작한다. ROUTINE은 루틴 편집, EXPLORE는 제품 탐색으로 이어진다.
      operationId: completeOnboarding
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CompleteOnboardingRequest' }
      responses:
        '200': { description: 온보딩 완료, content: { application/json: { schema: { $ref: '#/components/schemas/OnboardingResult' } } } }
        '401': { $ref: '#/components/responses/Problem' }
        '422': { $ref: '#/components/responses/Problem' }
...
  /products:
    get:
      tags: [Catalog]
      summary: 브랜드·제품명·종류로 제품을 커서 검색한다
      description: 보유 제품을 먼저, 그 안에서 제품 ID 순으로 반환한다. nextCursor는 클라이언트가 해석하지 않고 다음 요청에 그대로 보낸다.
      operationId: searchProducts
      parameters:
        - { name: query, in: query, schema: { type: string, maxLength: 100, default: '' } }
        - { name: cursor, in: query, schema: { type: [string, 'null'] } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 50, default: 24 } }
      responses:
        '200': { description: 검색 결과 한 페이지, content: { application/json: { schema: { $ref: '#/components/schemas/ProductPage' } } } }
        '422': { $ref: '#/components/responses/Problem' }
  /products/{productId}:
    get:
      tags: [Catalog]
      summary: 확인된 제품 정보를 조회한다
      operationId: getProduct
      parameters: [ { $ref: '#/components/parameters/ProductId' } ]
      responses:
        '200': { description: 제품, content: { application/json: { schema: { $ref: '#/components/schemas/Product' } } } }
        '404': { $ref: '#/components/responses/Problem' }
  /me/products:
    get:
      tags: [Collection]
      summary: 내 화장품을 조회한다
      operationId: listMyProducts
      responses:
        '200': { description: 내 화장품, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/UserProduct' } } } } }
    post:
      tags: [Collection]
      summary: 카탈로그 제품 또는 직접 입력 제품을 담는다
      operationId: addMyProduct
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddUserProductRequest' }
      responses:
        '201': { description: 담김, content: { application/json: { schema: { $ref: '#/components/schemas/UserProduct' } } } }
        '422': { $ref: '#/components/responses/Problem' }
  /me/products/{userProductId}:
    get:
      tags: [Collection]
      summary: 내 화장품 하나를 조회한다
      operationId: getMyProduct
      parameters: [ { $ref: '#/components/parameters/UserProductId' } ]
      responses:
        '200': { description: 내 화장품, content: { application/json: { schema: { $ref: '#/components/schemas/UserProduct' } } } }
        '404': { $ref: '#/components/responses/Problem' }
  /me/routines/current:
    get:
      tags: [Routines]
      summary: 현재 실제 사용 루틴을 조회한다
      operationId: getCurrentRoutine
      responses:
        '200': { description: 현재 루틴, content: { application/json: { schema: { $ref: '#/components/schemas/Routine' } } } }
        '404': { $ref: '#/components/responses/Problem' }
    put:
      tags: [Routines]
      summary: 제품과 순서를 저장하고 새 사용 경험을 시작한다
      operationId: replaceCurrentRoutine
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateRoutineRequest' }
      responses:
        '200': { description: 새 루틴 경험, content: { application/json: { schema: { $ref: '#/components/schemas/Experience' } } } }
        '409': { $ref: '#/components/responses/Problem' }
  /me/routines/baseline:
    get:
      tags: [Routines]
      summary: Rescue 비교에 쓰는 마지막 문제없던 루틴을 조회한다
      operationId: getBaselineRoutine
      responses:
        '200': { description: 비교 기준 루틴, content: { application/json: { schema: { $ref: '#/components/schemas/Routine' } } } }
        '404': { $ref: '#/components/responses/Problem' }
  /me/experiences/current:
    get:
      tags: [Experiences]
      summary: 현재 사용 중인 경험을 조회한다
      operationId: getCurrentExperience
      responses:
        '200': { description: 현재 경험, content: { application/json: { schema: { $ref: '#/components/schemas/Experience' } } } }
        '204': { description: 현재 경험 없음 }
  /me/experiences:
    post:
      tags: [Experiences]
      summary: 제품 하나 또는 현재 루틴에 제품을 더한 경험을 시작한다
      operationId: startExperience
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/StartExperienceRequest' }
      responses:
        '201': { description: 시작됨, content: { application/json: { schema: { $ref: '#/components/schemas/Experience' } } } }
  /me/experiences/{experienceId}:
    get:
      tags: [Experiences]
      summary: 사용 경험을 조회한다
      operationId: getExperience
      parameters: [ { $ref: '#/components/parameters/ExperienceId' } ]
      responses:
        '200': { description: 경험, content: { application/json: { schema: { $ref: '#/components/schemas/Experience' } } } }
        '404': { $ref: '#/components/responses/Problem' }
  /me/experiences/{experienceId}/records:
    post:
      tags: [Experiences]
      summary: 지금 또는 7일 시점의 사용 경험을 남긴다
      operationId: recordExperience
      parameters: [ { $ref: '#/components/parameters/ExperienceId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RecordExperienceRequest' }
      responses:
        '201': { description: 기록됨, content: { application/json: { schema: { $ref: '#/components/schemas/SavedExperienceRecord' } } } }
  /me/experiences/{experienceId}/complete:
    post:
      tags: [Experiences]
      summary: 7일 전이라도 현재 경험을 마친다
      operationId: completeExperience
      parameters: [ { $ref: '#/components/parameters/ExperienceId' } ]
      responses:
        '200': { description: 완료, content: { application/json: { schema: { $ref: '#/components/schemas/ApiMessage' } } } }
...
  /me/patterns/{patternId}:
    get:
      tags: [Patterns]
      summary: 패턴과 지지·반대 원본 기록을 조회한다
      operationId: getPattern
      parameters: [ { $ref: '#/components/parameters/PatternId' } ]
      responses:
        '200': { description: 패턴, content: { application/json: { schema: { $ref: '#/components/schemas/Pattern' } } } }
        '404': { $ref: '#/components/responses/Problem' }
  /ai/conversations:
    get:
      tags: [AI]
      summary: 내 AI 대화 목록을 조회한다
      operationId: listConversations
      responses:
        '200': { description: 대화 목록, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Conversation' } } } } }
    post:
      tags: [AI]
      summary: 공통 AI 채팅을 맥락과 초기 메시지로 시작한다
      operationId: createConversation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateConversationRequest' }
      responses:
        '201': { description: 첫 AI 답변 포함, content: { application/json: { schema: { $ref: '#/components/schemas/Conversation' } } } }
  /ai/conversations/{conversationId}:
    get:
      tags: [AI]
      summary: 대화 전체를 조회한다
      operationId: getConversation
      parameters: [ { $ref: '#/components/parameters/ConversationId' } ]
      responses:
        '200': { description: 대화, content: { application/json: { schema: { $ref: '#/components/schemas/Conversation' } } } }
        '404': { $ref: '#/components/responses/Problem' }
  /ai/conversations/{conversationId}/messages:
    post:
      tags: [AI]
      summary: 메시지를 보내고 동적 추천 입력이 포함된 AI 답변을 받는다
      operationId: sendConversationMessage
      parameters: [ { $ref: '#/components/parameters/ConversationId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SendMessageRequest' }
      responses:
        '200': { description: 갱신된 대화, content: { application/json: { schema: { $ref: '#/components/schemas/Conversation' } } } }
  /ai/conversations/{conversationId}/rescue/apply:
    post:
      tags: [AI, Routines]
      summary: 사용자가 승인한 Rescue 루틴을 새 경험으로 적용한다
      operationId: applyRescuePlan
      parameters: [ { $ref: '#/components/parameters/ConversationId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ClientRequest' }
      responses:
        '200': { description: 새 독립 루틴 경험, content: { application/json: { schema: { $ref: '#/components/schemas/Experience' } } } }
        '409': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
  parameters:
    ProductId: { name: productId, in: path, required: true, schema: { type: integer, format: int64 } }
    UserProductId: { name: userProductId, in: path, required: true, schema: { type: integer, format: int64 } }
    ExperienceId: { name: experienceId, in: path, required: true, schema: { type: integer, format: int64 } }
    PatternId: { name: patternId, in: path, required: true, schema: { type: integer, format: int64 } }
    ConversationId: { name: conversationId, in: path, required: true, schema: { type: integer, format: int64 } }
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 1 } }
  responses:
    Problem:
      description: RFC 9457 문제 응답
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    SignUpRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, pattern: '^[a-z0-9_]{4,24}$' }
        password: { type: string, minLength: 8, maxLength: 72, writeOnly: true }
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string, writeOnly: true }
    Auth:
      type: object
      required: [userId, username, displayName, demo, onboardingCompleted]
      properties:
        userId: { type: integer, format: int64 }
        username: { type: string }
        displayName: { type: string }
        demo: { type: boolean }
        onboardingCompleted: { type: boolean }
    QuickAccount:
      type: object
      required: [username, displayName]
      properties:
        username: { type: string }
        displayName: { type: string }
    CompleteOnboardingRequest:
      type: object
      required: [productIds, entryChoice, clientRequestId]
      properties:
        productIds:
          type: array
          maxItems: 8
          uniqueItems: true
          items: { type: integer, format: int64 }
        entryChoice: { type: string, enum: [PRODUCT, ROUTINE, EXPLORE] }
        focusProductId: { type: [integer, 'null'], format: int64 }
        clientRequestId: { type: string, minLength: 1 }
    OnboardingResult:
      type: object
      required: [user]
      properties:
        user: { $ref: '#/components/schemas/Auth' }
        experience: { anyOf: [ { $ref: '#/components/schemas/Experience' }, { type: 'null' } ] }
    Product:
      type: object
      required: [id, brand, name, category, verified, guide, facts, personalRecordCount, owned]
      properties:
        id: { type: integer, format: int64 }
        brand: { type: string }
        name: { type: string }
        category: { type: string }
        volume: { type: [string, 'null'] }
        versionLabel: { type: [string, 'null'] }
        verified: { type: boolean, description: legacy 확인 flag와 source-backed fact가 모두 있을 때만 true }
        guide: { $ref: '#/components/schemas/ProductGuide' }
        facts: { type: array, items: { $ref: '#/components/schemas/ProductFact' } }
        personalRecordCount: { type: integer, minimum: 0 }
        owned: { type: boolean }
        imageUrl: { type: [string, 'null'], format: uri }
    ProductGuide:
      type: object
      required: [summary, routineStep, usageType, usageTiming, usageInstructions, highlights, origin, generatedAt]
      properties:
        summary: { type: string, minLength: 1, maxLength: 500, description: 제품명과 등록 제형을 포함해 제품 종류를 설명하는 짧은 안내 }
        routineStep: { type: string, minLength: 1, maxLength: 100, description: category를 바탕으로 한 일반적인 루틴 위치 }
        usageType: { type: string, minLength: 1, maxLength: 100, description: 씻어내기·붙이기·남겨두기 등 일반 사용 형태 }
        usageTiming:
          type: array
          maxItems: 4
          items: { type: string, minLength: 1, maxLength: 100 }
        usageInstructions:
          type: array
          description: 제품 표시를 우선하도록 안내하는 일반 사용법. 출처 확인 제품 사실이 아님
          maxItems: 4
          items: { type: string, minLength: 1, maxLength: 300 }
        highlights:
          type: array
          description: 등록 제형을 첫 항목으로 포함해 제품 종류와 사용 형태를 이해하기 위한 특징 설명. 관찰·기록 과제가 아님
          maxItems: 4
          items: { $ref: '#/components/schemas/ProductHighlight' }
        origin: { type: string, enum: [AI_GENERATED, EDITORIAL] }
        generatedAt: { type: string, format: date-time }
    ProductHighlight:
      type: object
      required: [title, detail]
      properties:
        title: { type: string, minLength: 1, maxLength: 80 }
        detail: { type: string, minLength: 1, maxLength: 300 }
    ProductFact:
      type: object
      required: [type, text, sourceLabel, sourceUrl, checkedAt]
      properties:
        type:
          type: string
          enum: [DIRECTIONS, TEXTURE, LABEL_CLAIM, CAUTION, SUN_PROTECTION, INGREDIENT_LABEL, CERTIFICATION]
        text: { type: string, minLength: 1, maxLength: 500 }
        sourceLabel: { type: string, minLength: 1, maxLength: 150 }
        sourceUrl: { type: string, format: uri }
        checkedAt: { type: string, format: date-time }
    ProductPage:
      type: object
      required: [items, hasMore]
      properties:
        items: { type: array, maxItems: 50, items: { $ref: '#/components/schemas/Product' } }
        nextCursor: { type: [string, 'null'] }
        hasMore: { type: boolean }
    UserProduct:
      type: object
      required: [id, addedAt]
      properties:
        id: { type: integer, format: int64 }
        product: { anyOf: [ { $ref: '#/components/schemas/Product' }, { type: 'null' } ] }
        customBrand: { type: [string, 'null'] }
        customName: { type: [string, 'null'] }
        customCategory: { type: [string, 'null'] }
        memo: { type: [string, 'null'] }
        addedAt: { type: string }
    AddUserProductRequest:
      type: object
      description: productId 또는 customName 중 하나를 사용한다.
      properties:
        productId: { type: [integer, 'null'], format: int64 }
        customBrand: { type: [string, 'null'] }
        customName: { type: [string, 'null'] }
        customCategory: { type: [string, 'null'] }
        memo: { type: [string, 'null'] }
    RoutineItem:
      type: object
      required: [userProductId, productName, brand, category, timeSlot, position, frequency]
      properties:
        userProductId: { type: integer, format: int64 }
        productName: { type: string }
        brand: { type: string }
        category: { type: string }
        timeSlot: { type: string, enum: [MORNING, EVENING, BOTH] }
        position: { type: integer, minimum: 0 }
        frequency: { type: string }
    RoutineItemInput:
      type: object
      required: [userProductId, timeSlot, frequency]
      properties:
        userProductId: { type: integer, format: int64 }
        timeSlot: { type: string, enum: [MORNING, EVENING, BOTH] }
        frequency: { type: string, maxLength: 20 }
    Routine:
      type: object
      required: [id, name, dayPart, status, startedAt, items]
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        dayPart: { type: string, enum: [MORNING, EVENING, ANYTIME] }
        status: { type: string, enum: [CURRENT, BASELINE, PAST] }
        startedAt: { type: string }
        items: { type: array, items: { $ref: '#/components/schemas/RoutineItem' } }
    UpdateRoutineRequest:
      type: object
      required: [name, items]
      properties:
        name: { type: string, minLength: 1 }
        items: { type: array, minItems: 1, maxItems: 12, items: { $ref: '#/components/schemas/RoutineItemInput' } }
    StartExperienceRequest:
      type: object
      required: [userProductId, mode, clientRequestId]
      properties:
        userProductId: { type: integer, format: int64 }
        mode: { type: string, enum: [PRODUCT, ROUTINE] }
        dayPart: { type: string, enum: [MORNING, EVENING, ANYTIME], default: EVENING }
        clientRequestId: { type: string, minLength: 1 }
    Experience:
      type: object
      required: [id, subjectType, title, subtitle, status, startedAt, reviewDueAt, day, daysUntilReview, reviewDue]
      properties:
        id: { type: integer, format: int64 }
        subjectType: { type: string, enum: [PRODUCT, ROUTINE] }
        routineId: { type: [integer, 'null'], format: int64 }
        userProductId: { type: [integer, 'null'], format: int64 }
        title: { type: string }
        subtitle: { type: string }
        status: { type: string, enum: [ACTIVE, COMPLETED, CANCELLED] }
        startedAt: { type: string }
        reviewDueAt: { type: string }
        day: { type: integer, minimum: 1 }
        daysUntilReview: { type: integer, minimum: 0 }
        reviewDue: { type: boolean }
        routine: { anyOf: [ { $ref: '#/components/schemas/Routine' }, { type: 'null' } ] }
        product: { anyOf: [ { $ref: '#/components/schemas/UserProduct' }, { type: 'null' } ] }
        latestRecord: { anyOf: [ { $ref: '#/components/schemas/ExperienceRecord' }, { type: 'null' } ] }
    RecordExperienceRequest:
      type: object
      required: [sentiment, discomfort, clientRequestId]
      properties:
        sentiment: { type: string, enum: [LIKED, DISAPPOINTED, UNSURE] }
        note: { type: string, maxLength: 1200 }
        tags: { type: array, maxItems: 8, items: { type: string } }
        discomfort: { type: string, enum: [NOT_REPORTED, REPORTED, UNKNOWN] }
        adherence: { type: string, enum: [MATCHED, PARTIAL, DIFFERENT, UNKNOWN], default: MATCHED }
        clientRequestId: { type: string, minLength: 1 }
    ExperienceRecord:
      type: object
      required: [id, productName, sentiment, note, discomfort, adherence, tags, createdAt]
      properties:
        id: { type: integer, format: int64 }
        sessionId: { type: [integer, 'null'], format: int64 }
        userProductId: { type: [integer, 'null'], format: int64 }
        productName: { type: string }
        sentiment: { type: string, enum: [LIKED, DISAPPOINTED, UNSURE] }
        note: { type: string }
        discomfort: { type: string, enum: [NOT_REPORTED, REPORTED, UNKNOWN] }
        adherence: { type: string, enum: [MATCHED, PARTIAL, DIFFERENT, UNKNOWN] }
        tags: { type: array, items: { type: string } }
        createdAt: { type: string }
    SavedExperienceRecord:
      type: object
      required: [record, rescueSuggested]
      properties:
        record: { $ref: '#/components/schemas/ExperienceRecord' }
        linkedPatternId: { type: [integer, 'null'], format: int64 }
        rescueSuggested: { type: boolean }
    PatternEvidence:
      type: object
      required: [recordId, productName, note, sentiment, polarity, createdAt]
      properties:
        recordId: { type: integer, format: int64 }
        productName: { type: string }
        note: { type: string }
        sentiment: { type: string }
        polarity: { type: string, enum: [SUPPORTS, CONTRADICTS] }
        createdAt: { type: string }
    Pattern:
      type: object
      required: [id, title, summary, confidenceNote, supportingCount, contradictingCount, evidence]
      properties:
        id: { type: integer, format: int64 }
        title: { type: string }
        summary: { type: string }
        confidenceNote: { type: string }
        supportingCount: { type: integer, minimum: 0 }
        contradictingCount: { type: integer, minimum: 0 }
        evidence: { type: array, items: { $ref: '#/components/schemas/PatternEvidence' } }
    Message:
      type: object
      required: [id, role, content, suggestedReplies, evidenceRefs, status, createdAt]
      properties:
        id: { type: integer, format: int64 }
        role: { type: string, enum: [USER, ASSISTANT] }
        content: { type: string }
        suggestedReplies: { type: array, maxItems: 3, items: { type: string, maxLength: 80 } }
        evidenceRefs: { type: array, maxItems: 8, items: { type: string, maxLength: 64 } }
        status: { type: string, enum: [READY, FALLBACK, FAILED] }
        createdAt: { type: string }
    RescuePlan:
      type: object
      required: [id, title, rationale, status]
      properties:
        id: { type: integer, format: int64 }
        baseRoutineId: { type: [integer, 'null'], format: int64 }
        title: { type: string }
        rationale: { type: string }
        removeUserProductId: { type: [integer, 'null'], format: int64 }
        removeProductName: { type: [string, 'null'] }
        status: { type: string, enum: [PROPOSED, APPLIED, DECLINED, BLOCKED] }
        appliedExperienceId: { type: [integer, 'null'], format: int64 }
    Conversation:
      type: object
      description: 공통 AI 대화. 일반 대화 중 불편 표현이 확인되면 서버가 같은 대화의 mode를 RESCUE로 전환할 수 있다.
      required: [id, mode, status, messages, quickReplies, safetyBoundary]
      properties:
        id: { type: integer, format: int64 }
        mode: { type: string, enum: [GENERAL, PRODUCT, RECOMMEND, PATTERN, RESCUE] }
        productId: { type: [integer, 'null'], format: int64 }
        experienceId: { type: [integer, 'null'], format: int64 }
        status: { type: string, enum: [OPEN, CLOSED] }
        messages: { type: array, items: { $ref: '#/components/schemas/Message' } }
        quickReplies: { type: array, maxItems: 3, items: { type: string } }
        rescuePlan: { anyOf: [ { $ref: '#/components/schemas/RescuePlan' }, { type: 'null' } ] }
        safetyBoundary: { type: boolean }
    CreateConversationRequest:
      type: object
      required: [mode, initialPrompt, clientRequestId]
      properties:
        mode: { type: string, enum: [GENERAL, PRODUCT, RECOMMEND, PATTERN, RESCUE], description: 요청한 진입 맥락. initialPrompt가 명백한 불편 표현이면 응답 mode는 RESCUE로 전환될 수 있다. }
        productId: { type: [integer, 'null'], format: int64 }
        experienceId: { type: [integer, 'null'], format: int64 }
        initialPrompt: { type: string, minLength: 1, maxLength: 1500 }
        clientRequestId: { type: string, minLength: 1 }
    SendMessageRequest:
      type: object
      required: [text, clientRequestId]
      properties:
        text: { type: string, minLength: 1, maxLength: 1500 }
        clientRequestId: { type: string, minLength: 1 }
    ClientRequest:
      type: object
      required: [clientRequestId]
      properties:
        clientRequestId: { type: string, minLength: 1 }
...
    ApiMessage:
      type: object
      required: [message]
      properties: { message: { type: string } }
    Problem:
      type: object
      required: [type, title, status, detail, code, retryable]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string }
        retryable: { type: boolean }